backend/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
idna==3.6
pillow==10.2.0
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
stripe==8.1.0
typing_extensions==4.8.0
urllib3==2.2.0
